rename config-loader.ts to configLoader.ts
